rag_project/rag_project/requirements.txt
openai>=1.30.0
streamlit>=1.35.0
faiss-cpu>=1.8.0
numpy>=1.26.0
PyPDF2>=3.0.0

# Optional — for cross-encoder re-ranking
# sentence-transformers>=2.7.0
